addded GetDefaultStressRequest + golang runtime test
