+++ pkg/processor/runtime/golang/test/golang_test.go
@@ -164,6 +164,14 @@ func (suite *TestSuite) TestOutputs() {
 	})
 }
 
+func (suite *TestSuite) TestStress(){
+
+	// Test stress with default stress-test configurations
+	stressTestResult := suite.BlastHTTP(suite.GetDefaultStressRequest())
+
+	suite.Require().Equal(true, stressTestResult )
+}
+
 func TestIntegrationSuite(t *testing.T) {
 	if testing.Short() {
 		return
